+++ android/src/main/java/dev/barooni/capacitor/calendar/implementation/CapacitorCalendar.kt
@@ -1,6 +1,8 @@
 package dev.barooni.capacitor.calendar.implementation
 
+import android.content.ContentValues
 import android.content.Intent
+import android.net.Uri
 import android.provider.CalendarContract
 import com.getcapacitor.PermissionState
 import com.getcapacitor.PluginCall
@@ -8,11 +10,15 @@ import dev.barooni.capacitor.calendar.CapacitorCalendarPlugin
 import dev.barooni.capacitor.calendar.PluginError
 import dev.barooni.capacitor.calendar.models.enums.CalendarPermissionScope
 import dev.barooni.capacitor.calendar.models.inputs.CheckPermissionInput
+import dev.barooni.capacitor.calendar.models.inputs.CreateEventInput
 import dev.barooni.capacitor.calendar.models.inputs.CreateEventWithPromptInput
+import dev.barooni.capacitor.calendar.models.inputs.ModifyEventWithPromptInput
 import dev.barooni.capacitor.calendar.models.inputs.RequestAllPermissionsInput
 import dev.barooni.capacitor.calendar.models.inputs.RequestPermissionInput
 import dev.barooni.capacitor.calendar.models.results.CheckAllPermissionsResult
 import dev.barooni.capacitor.calendar.models.results.CheckPermissionResult
+import dev.barooni.capacitor.calendar.models.results.CreateEventResult
+import dev.barooni.capacitor.calendar.utils.ImplementationHelper
 
 class CapacitorCalendarNew(
     private val plugin: CapacitorCalendarPlugin,
@@ -48,17 +54,39 @@ class CapacitorCalendarNew(
         input: CreateEventWithPromptInput,
         callback: (PluginCall, Intent, String) -> Unit,
     ) {
-        val intent =
-            input.intent
-                .setData(CalendarContract.Events.CONTENT_URI)
-                .putExtra(CalendarContract.Events.TITLE, input.title)
-                .putExtra(CalendarContract.EXTRA_EVENT_BEGIN_TIME, input.startDate)
-                .putExtra(CalendarContract.EXTRA_EVENT_END_TIME, input.endDate)
-                .putExtra(CalendarContract.EXTRA_EVENT_ALL_DAY, input.isAllDay)
-        input.location?.let { intent.putExtra(CalendarContract.Events.EVENT_LOCATION, it) }
-        input.description?.let { intent.putExtra(CalendarContract.Events.DESCRIPTION, it) }
-        input.availability?.let { intent.putExtra(CalendarContract.Events.AVAILABILITY, it) }
-        input.invitees?.let { intent.putExtra(Intent.EXTRA_EMAIL, it) }
-        callback(input.call, intent, input.callbackName)
+        callback(input.call, input.intent, input.callbackName)
+    }
+
+    fun modifyEventWithPrompt(
+        input: ModifyEventWithPromptInput,
+        callback: (PluginCall, Intent, String) -> Unit,
+    ) {
+        callback(input.call, input.intent, input.callbackName)
+    }
+
+    fun createEvent(input: CreateEventInput): CreateEventResult {
+        val cr = plugin.context.contentResolver
+        val calendarId: Long = input.calendarId ?: ImplementationHelper.getDefaultCalendarId(cr)
+
+        val values =
+            ContentValues().apply {
+                put(CalendarContract.Events.TITLE, input.title)
+                put(CalendarContract.Events.CALENDAR_ID, calendarId)
+                put(CalendarContract.Events.EVENT_TIMEZONE, input.timezoneId)
+                input.isAllDay?.let { put(CalendarContract.Events.ALL_DAY, it) }
+                input.location?.let { put(CalendarContract.Events.EVENT_LOCATION, it) }
+                input.startDate?.let { put(CalendarContract.Events.DTSTART, it) }
+                input.endDate?.let { put(CalendarContract.Events.DTEND, it) }
+                input.description?.let { put(CalendarContract.Events.DESCRIPTION, it) }
+                input.availability?.let { put(CalendarContract.Events.AVAILABILITY, it) }
+                input.organizer?.let { put(CalendarContract.Events.ORGANIZER, it) }
+                input.duration?.let { put(CalendarContract.Events.DURATION, it) }
+                input.color?.let { put(CalendarContract.Events.EVENT_COLOR, it) }
+            }
+        val uri: Uri? = cr.insert(CalendarContract.Events.CONTENT_URI, values)
+        val eventId: Long = uri?.lastPathSegment?.toLong() ?: throw PluginError.FailedToRetrieveEventId
+        input.attendees?.let { ImplementationHelper.insertAttendeesToEvent(eventId, cr, it) }
+        input.alerts?.let { ImplementationHelper.insertAlertsToEvents(eventId, cr, it) }
+        return CreateEventResult(eventId)
     }
 }