+++ android/src/main/java/dev/barooni/capacitor/calendar/models/inputs/CreateEventWithPromptInput.kt
@@ -9,13 +9,28 @@ data class CreateEventWithPromptInput(
     val call: PluginCall,
     val callbackName: String,
 ) {
-    val title: String = call.getString("title", "") ?: ""
+    private val title: String = call.getString("title", "") ?: ""
     val location: String? = call.getString("location")
-    val startDate: Long = ImplementationHelper.getCalendarFromTimestamp(call.getLong("startDate")).timeInMillis
-    val endDate: Long = ImplementationHelper.getCalendarFromTimestamp(call.getLong("endDate")).timeInMillis
-    val isAllDay: Boolean = call.getBoolean("isAllDay") ?: false
+    val startDate: Long? = call.getLong("startDate")?.let { ImplementationHelper.getCalendarFromTimestamp(it).timeInMillis }
+    val endDate: Long? = call.getLong("endDate")?.let { ImplementationHelper.getCalendarFromTimestamp(it).timeInMillis }
+    val isAllDay: Boolean? = call.getBoolean("isAllDay")
     val description: String? = call.getString("description")
     val availability: Int? = call.getInt("availability")
     val invitees: String? = ImplementationHelper.jsArrayToComaSeparatedString(call.getArray("invitees"))
     val intent: Intent = Intent(Intent.ACTION_INSERT).setData(CalendarContract.Events.CONTENT_URI)
+
+    init {
+        setEventData()
+    }
+
+    private fun setEventData() {
+        intent.putExtra(CalendarContract.Events.TITLE, title)
+        location?.let { intent.putExtra(CalendarContract.Events.EVENT_LOCATION, it) }
+        startDate?.let { intent.putExtra(CalendarContract.EXTRA_EVENT_BEGIN_TIME, it) }
+        endDate?.let { intent.putExtra(CalendarContract.EXTRA_EVENT_END_TIME, it) }
+        isAllDay?.let { intent.putExtra(CalendarContract.EXTRA_EVENT_ALL_DAY, it) }
+        description?.let { intent.putExtra(CalendarContract.Events.DESCRIPTION, it) }
+        availability?.let { intent.putExtra(CalendarContract.Events.AVAILABILITY, it) }
+        invitees?.let { intent.putExtra(Intent.EXTRA_EMAIL, it) }
+    }
 }