+++ android/src/main/java/dev/barooni/capacitor/calendar/models/results/ModifyEventWithPromptResult.kt
@@ -3,10 +3,10 @@ package dev.barooni.capacitor.calendar.models.results
 import com.getcapacitor.JSObject
 import dev.barooni.capacitor.calendar.models.templates.JSResult
 
-class CreateEventWithPromptResult : JSResult {
+class ModifyEventWithPromptResult : JSResult {
     override fun toJSON(): JSObject {
         val result = JSObject()
-        result.put("id", null)
+        result.put("result", null)
         return result
     }
 }