+++ ios/Sources/CapacitorCalendarPlugin/Models/Results/CreateEventResult.swift
@@ -1,0 +1,18 @@
+import Capacitor
+
+struct CreateEventResult: JSResult {
+    let id: String
+
+    init(id: String?) throws {
+        guard let id = id else {
+            throw PluginError.failedToRetrieveEventId
+        }
+        self.id = id
+    }
+
+    func toJSON() -> JSObject {
+        var result = JSObject()
+        result["id"] = id
+        return result
+    }
+}