+++ src/sub-definitions/calendar-access.ts
@@ -1,5 +1,5 @@
 import { CalendarPermissionScope } from "../schemas/enums/calendar-permission-scope";
-import { PermissionState } from "@capacitor/core";
+import type { PermissionState } from "@capacitor/core";
 
 /**
  * @since 7.1.0