+++ src/sub-definitions/calendar-operations.ts
@@ -1,0 +1,10 @@
+export interface CalendarOperations {
+  /**
+   * TODO: Throw unimplemented error on Android
+   * Save the changes to the calendar.
+   *
+   * @platform iOS
+   * @since 7.1.0
+   */
+  commit(): Promise<void>;
+}