+++ src/sub-definitions/reminders-access.ts
@@ -1,4 +1,4 @@
-import { PermissionState } from "@capacitor/core";
+import type { PermissionState } from "@capacitor/core";
 
 /**
  * @since 7.1.0
@@ -7,9 +7,6 @@ export interface RemindersAccess {
   /**
    * Requests read and write access to the reminders.
    *
-   * @see {@link CalendarPermissionScope}
-   * @since 5.4.0
-   *
    * @permissions
    * | Platform  | Required |
    * |-----------|---------------------|
@@ -17,6 +14,8 @@ export interface RemindersAccess {
    * | iOS 10-16 | `NSRemindersUsageDescription` |
    *
    * @platform iOS
+   * @see {@link CalendarPermissionScope}
+   * @since 5.4.0
    */
   requestFullRemindersAccess(): Promise<{ result: PermissionState }>;
 }